EXTERNAL_REVIEW_2026_04 cleanup items 8-14, 16-23
Dead code deletion:
  - storage/batch_iterator.go (511 lines, zero callers outside the
    file; contained a stale pre-expansion index enum and
    natural-order Tx encoding). Review gave the path as executor/
    but the file lived in storage/.
  - executor/context_minimal.go (9-line TODO stub; BaseContext
    already provides the no-op behavior).
  - executor/pattern_match.go: MemoryPatternMatcher struct and its
    four methods (unreachable — NewMemoryPatternMatcher returns an
    IndexedMemoryMatcher). Shared pattern-match routines below are
    kept because indexed_memory_matcher.go and test_fixtures.go
    still call them.
  - query/tuple_builder.go and tuple_builder_optimized.go (both
    self-documented "CANDIDATE FOR REMOVAL"; only production user
    was InternedTupleBuilder). storage/tuple_builder_bench_test.go
    deleted with them — it only compared the now-removed variants.
  - storage/utils.go: concatBytes moved into key_encoder_base.go
    (its natural home — called by every key encoder). File name
    "utils.go" violated the project naming rule.
  - docs/STREAMING_ARCHITECTURE_DECISION.md: the "pragmatic
    compromise" it documented (retain streaming globals as
    effectively read-only) has since been fully reversed; all
    streaming flags now thread through ExecutorOptions.

Test relocation:
  - executor/datom_relation.go renamed to datom_relation_test.go.
    NewDatomIterator/NewDatomRelation only have test callers now.
  - storage/bug5_vector_avet_test.go gains a local encodeKeyword
    (was defined in the deleted batch_iterator.go; nothing outside
    that test used it).

Correctness / hygiene:
  - annotations/output.go: isTerminal now uses
    github.com/mattn/go-isatty instead of an fd==1||fd==2 check
    that incorrectly emitted ANSI color codes into files
    redirected from stdout/stderr.
  - executor/aggregation.go: unconditional fmt.Printf("AGGREGATE
    BUG: ...") on a missing-symbol edge case removed — fired to
    stdout in production.
  - executor/predicate_classifier.go:
    SplitPredicatesForPattern now shallow-copies phase before
    replacing Predicates, so a caller passing a shared
    *planner.Phase no longer sees its Predicates slice stomped.

Enum consistency:
  - planner/types.go IndexType now lists all seven storage
    indices (EAVT, EATV, AEVT, AETV, AVET, VAET, TAEV) with a
    note that selectIndexForMask does not yet emit EATV or AETV.
    The prior five-value copy was left stale by the CRDT-Tx
    expansion. Integer values are local to the planner and must
    not be cross-cast to storage's enum.
  - storage/hash_join_matcher.go: stale "// 3" / "// 4" case
    comments (numerically wrong after EATV/AETV expansion)
    removed; AETV comment kept since it clarifies the CRDT
    ordering.

Doc drift:
  - README.md: Datomic parity figure reconciled to ~80% to match
    DATOMIC_COMPATIBILITY.md / CLAUDE.md / ARCHITECTURE.md.
  - executor/iterator_composition.go: stale "managed by
    ExecutorOptions but kept for backward compatibility" comment
    removed — the referenced variables no longer exist.
  - docs/decisions/README.md: link to the deleted
    STREAMING_ARCHITECTURE_DECISION.md removed.

go.mod: mattn/go-isatty promoted from indirect to direct.

Item 15 (RepeatOffsets wire-or-delete) deferred — it has
working unit tests and represents a half-wired compression
feature, not dead code.
Item 24 is moot (key_mask_iterator.go was already deleted
in commit 03205b0).

Net diff: 1,254 lines deleted, 76 added. Full suite green.
