+++ README.md
@@ -915,7 +915,7 @@ See [docs/papers/README.md](docs/papers/README.md) for complete details.
 
 - **[docs/reference/QUERY_BUILDER.md](docs/reference/QUERY_BUILDER.md)** - Go-native query builder (recommended)
 - **[ARCHITECTURE.md](ARCHITECTURE.md)** - System architecture and design decisions
-- **[DATOMIC_COMPATIBILITY.md](DATOMIC_COMPATIBILITY.md)** - Feature comparison (~70% compatibility)
+- **[DATOMIC_COMPATIBILITY.md](DATOMIC_COMPATIBILITY.md)** - Feature comparison (~80% compatibility)
 - **[DOCUMENTATION_INDEX.md](DOCUMENTATION_INDEX.md)** - Complete documentation guide
 
 ### Deep Dives
@@ -978,7 +978,7 @@ We welcome contributions! Here's how to get started:
 
 ## Datomic Compatibility
 
-Janus implements **~70% of Datomic's feature set** (weighted by typical usage), focusing on the most commonly used features:
+Janus implements **~80% of Datomic's feature set** (weighted by typical usage), focusing on the most commonly used features:
 
 **Implemented:**
 - Core queries: Patterns, joins, variables
