--- docs/STREAMING_ARCHITECTURE_DECISION.md
@@ -1,105 +1,0 @@
-# Streaming Architecture Decision
-
-**Date**: October 2025
-**Status**: Implemented
-
-## Context
-
-We identified that the Datalog query engine was forcing materialization of intermediate results even though it had streaming relation support. This was causing:
-- **Memory explosion**: Intermediate results fully materialized
-- **Poor performance**: 1.5-2.5x slower than necessary
-- **Wasted resources**: 50-99% more memory than needed
-
-## Solution Implemented
-
-We implemented a complete streaming preservation pipeline:
-
-### 1. Iterator Composition
-- `FilterIterator`: Lazy filtering without materialization
-- `ProjectIterator`: Lazy projection to subset of symbols
-- `TransformIterator`: Lazy tuple transformation
-- `PredicateFilterIterator`: Lazy predicate evaluation
-- `FunctionEvaluatorIterator`: Lazy function evaluation
-
-### 2. Buffered Iterators
-Solves the single-consumption problem of streaming iterators:
-- Buffers tuples on first iteration
-- Allows multiple consumers to iterate independently
-- Maintains streaming semantics while supporting re-iteration
-
-### 3. Symmetric Hash Joins
-For true streaming-to-streaming joins:
-- Dual hash tables for incremental processing
-- Processes tuples as they arrive
-- No need to materialize either side
-
-## Configuration Approach
-
-### The Problem with Global Variables
-Initially, we introduced global configuration variables (`EnableIteratorComposition`, `EnableTrueStreaming`, etc.) which was identified as an anti-pattern. The user correctly pointed out this was bad practice.
-
-### The Ideal Solution (Not Implemented)
-Thread `ExecutorOptions` through all relation constructors and operations. This would require:
-- Modifying 50+ function signatures
-- Passing options through entire call chain
-- Massive refactoring across codebase
-
-### The Pragmatic Compromise (Implemented)
-We kept global variables BUT made them effectively read-only:
-1. **Initialize once**: Set from `PlannerOptions` when executor is created
-2. **Never modify**: Documented as READ-ONLY after initialization
-3. **Centralized control**: Only `initializeGlobalOptions()` modifies them
-4. **TODO documented**: Clear path to future refactoring
-
-```go
-// Global configuration variables set once from PlannerOptions when executor is created
-// These are READ-ONLY after initialization - do not modify directly!
-// TODO: Future refactoring should pass options through relation constructors
-var (
-    EnableIteratorComposition = true  // Lazy evaluation
-    EnableTrueStreaming      = true  // No auto-materialization
-    EnableSymmetricHashJoin  = false // Conservative for now
-    // ...
-)
-```
-
-## Why This Approach?
-
-1. **Avoids massive refactoring**: No need to change 50+ function signatures
-2. **Maintains encapsulation**: Options flow through PlannerOptions → Executor
-3. **Clear ownership**: Only executor initialization sets globals
-4. **Safe for production**: No random mutation of global state
-5. **Clear upgrade path**: TODO documents future refactoring
-
-## Performance Results
-
-With streaming enabled by default:
-- **1.5-2.5× faster** query execution
-- **50-99% less memory** for selective queries
-- **Zero configuration** needed (enabled by default)
-
-## Testing Strategy
-
-Tests should use `NewExecutorWithOptions` to control configuration, not modify globals directly:
-
-```go
-// GOOD: Use executor options
-opts := planner.PlannerOptions{
-    EnableIteratorComposition: true,
-    EnableTrueStreaming: true,
-}
-exec := NewExecutorWithOptions(matcher, opts)
-
-// BAD: Don't modify globals in tests
-EnableIteratorComposition = false  // Don't do this!
-```
-
-## Future Work
-
-Eventually refactor to pass `ExecutorOptions` through relation operations:
-1. Add `opts ExecutorOptions` parameter to relation constructors
-2. Thread options through join/filter/project operations
-3. Remove global variables entirely
-4. Update all tests to pass options explicitly
-
-But this is not urgent - the current approach is safe and functional.
